Split sync chat apply helpers
